Address code review feedback: extract ColourProfileSection, fix stale closure, remove redundant state, add error toast, expose defaults from backend
- Extract colour profile UI and logic into ColourProfileSection to bring Profile.tsx under 200 lines
- Backend exposes defaultColourProfile as a read-only serializer field, eliminating the duplicate DEFAULT_COLOUR_PROFILE constant in the frontend
- Remove colourProfileEnabled local state and its useEffect; use query cache directly for optimistic toggle with rollback
- Fix stale closure in debounced colour save rollback by keeping savedColoursRef in sync on each render
- Add toast.error when the colour-blind mode toggle fails
- Add Suspense requirement comment to useCustomNationColours
